Put SIM800L usage example into a separate file
In order to lower the memory footprint of this module, let's
deliberately move the usage example code into a separate file.
